requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.3
colorama==0.4.4
DateTime==4.4
Flask==2.0.2
idna==3.3
importlib-metadata==4.10.1
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
python-dateutil==2.8.2
pytz==2021.3
requests==2.27.1
six==1.16.0
typing_extensions==4.0.1
urllib3==1.26.8
Werkzeug==2.0.2
zipp==3.7.0
zope.interface==5.4.0
